Increase database timeouts for large data volume
We are inserting hundreds of thousands of events the first time this
is run. The previous 45-second timeout was not long enough to perform
the insertion.

A better approach is to batch up the inserts into a thousand or so at
a time. This commit does not do that because this is only an
experimental project.