Stringify GUIDs retrieved from the database
Some Audit Events do not include Org or Space GUIDs. For these we
set the database field to `null::uuid` on lines 66–102 of
`db/store.go`.

When retrieving data from the database, I discovered that `rows.Scan`
was failing for these events. It could not handle `cfclient.Event`'s'
`OrganizationGUID` or `SpaceGUID` fields being null, because they are
string fields.

We cannot fix this in Go without changing the `cfclient.Event` struct,
which I do not want to do. This commit works around the problem by
casting the GUID fields to text when retrieving them. GUID fields set
to `null::uuid` appear to come back as an all-zero GUID.