Fetch events since since last event or else four weeks ago
This commit implements logic for which CC Audit Events to fetch. The
only control we have around what the API endpoints return is saying
what time to return events from.

If there are events in the database, we want to request events since
the most recently stored one. To ensure we don't miss events at
similar times, we request events from 5 minutes before then.

Events each have GUIDs and will be deduplicated by
`PRIMARY KEY (guid)` on the `cf_audit_events` table.

If there are no events in the database, we go back four weeks
(28 days.) Cloud Controller generally stores 31 days, but 28 days is
what I happened to code. It's possible there will not be a full 28
days of data in CC and that will be OK.