Insert events to the database page-by-page
Before this commit we fetched Audit Events from Cloud Controller using
the `cfclient.ListEventsByQuery` function. This function fetches all
the pages and then returns all the events. There are a few downsides:

  - This function fetched all events into memory and then we saved all
    events at once. On GOV.UK PaaS, it takes tens of thousands of API
    requests to fetch all pages. Just one HTTP or DB error during this
    process would waste all that work.

  - The function did not support ratelimiting, and the API call seems
    relatively expensive.

These issues are both quite important to being production-ready. It
is probably easier to move the fetching into our code than patch both
of them in `cfclient`. We had to use channels to pass things around
and that would be a major departure from the rest of the `cfclient`
library.