Add JSONB metadata field to the audit events table
A column type of JSONB is chosen for flexibility; the alternative
would be hstore. This is done as an ALTER TABLE so that the SQL will
apply on top of existing databases cleanly.

`ADD COLUMN IF NOT EXISTS` was new in Postgres 9.6, which is OK
because we are using Postgres 10 for this.